Fix JDBC date time parsing issue
